test(router-outlet): migrate
